// src/downloader/compiler/types.ts
export interface IDist {
  type: string;
  name: string;
  version: string;
  platform: string;
  dev: boolean;
}

export interface IDistFile {
  dist: IDist;
  path: string;
}

export interface IDownloadCompilerOptions {
  path: string;
  version: string;
  dev: boolean;
  dists: string[];
}

export interface IDownloadOpts {
  url: string;
  path: string;
}